requirements.txt
toml==0.10.2
PyMuPDF>=1.23.0
Pillow>=9.0.0
